+++ service/java/io/pedestal/servlet/mock/MockAsyncContext.java
@@ -1,0 +1,88 @@
+// Copyright 2025 Nubank NA
+//
+// The use and distribution terms for this software are covered by the
+// Eclipse Public License 1.0 (http://opensource.org/licenses/eclipse-1.0)
+// which can be found in the file epl-v10.html at the root of this distribution.
+//
+// By using this software in any fashion, you are agreeing to be bound by
+// the terms of this license.
+//
+// You must not remove this notice, or any other, from this software.
+
+package io.pedestal.servlet.mock;
+
+import jakarta.servlet.*;
+
+/** @since 0.8.0 */
+class MockAsyncContext implements AsyncContext {
+    private final MockState state;
+
+    MockAsyncContext(MockState state) {
+        this.state = state;
+    }
+
+    @Override
+    public ServletRequest getRequest() {
+        return state.request;
+    }
+
+    @Override
+    public ServletResponse getResponse() {
+        return state.response;
+    }
+
+    @Override
+    public boolean hasOriginalRequestAndResponse() {
+        return false;
+    }
+
+    @Override
+    public void dispatch() {
+
+    }
+
+    @Override
+    public void dispatch(String path) {
+
+    }
+
+    @Override
+    public void dispatch(ServletContext context, String path) {
+
+    }
+
+    @Override
+    public void complete() {
+    state.complete();
+    }
+
+    @Override
+    public void start(Runnable run) {
+
+    }
+
+    @Override
+    public void addListener(AsyncListener listener) {
+
+    }
+
+    @Override
+    public void addListener(AsyncListener listener, ServletRequest servletRequest, ServletResponse servletResponse) {
+
+    }
+
+    @Override
+    public <T extends AsyncListener> T createListener(Class<T> clazz) throws ServletException {
+        return null;
+    }
+
+    @Override
+    public void setTimeout(long timeout) {
+
+    }
+
+    @Override
+    public long getTimeout() {
+        return 0;
+    }
+}